[Fixed] Erro do Format Price

+++ Gateway/Config/Config.php
@@ -87,9 +87,9 @@ class Config extends PaymentConfig
      *
      * @param float $amount
      *
-     * @return int
+     * @return float
      */
-    public function formatPrice($amount): int
+    public function formatPrice($amount): float
     {
         return $amount * self::ROUND_UP;
     }
